ArtASCII

Description

Contém funções relacionadas com o desenho dos caracteres no programa Draw

Documentation

type Art = [String]

Representa um tile em ArtASCII

artB :: Art

O Art do Tile do Tiletype B

artC :: Art

O Art do Tile do Tiletype C

artE :: Art

O Art do Tile do Tiletype E

artN :: Art

O Art do Tile do Tiletype N

artVoid :: Art

Um Art que representa um espaço vazio

buildMap :: [Tile] -> Limits -> Map

Organiza uma matriz de Maybe Tile

drawMap :: Map -> String

Converte os tiles do mapa para Arts e depois transforma-os numa string

artTiles2String :: [[Art]] -> String

Constroi uma String com a matriz de Art

tile2Art :: Maybe Tile -> Art

drawMeeple :: Art -> Maybe Meeple -> Art

coord :: Location -> Char -> Location

coordInv :: Location -> Char -> Location

rotateArt :: Art -> Char -> Art

getMatrixValue :: [[a]] -> Location -> a