Escritor

Description

Lê as informações do XML para as estruturas do Haskassone.

Documentation

tile2xmlString :: Tile -> String

Converte um objecto Tile no seu equivalente em XML

meeple2xmlString :: Meeple -> String

Converte um objecto Meeple no seu equivalente em XML

string2qname :: String -> QName

Obtém um QName dada uma String

elements2contents :: [Element] -> [Content]

Obtém uma lista de Contents a partir de uma lista de Elements

player2element :: Player -> Element

Converte um Player num Element

next2element :: Next -> Element

Converte um Next num Element

meeple2element :: Meeple -> Element

Converte um Meeple num Element

tile2element :: Tile -> Element

Converte um Tile num Element

board2element :: Board -> Element

Converte um Board num Element | Se o TileType do Next for - então não deve ser escrito um Next