Leitor

Description

Lê as informações do XML para as estruturas do Haskassone.

Documentation

data Player

Um jogador. O jogador é identificado por um valor inteiro e tem uma pontuação.

Constructors

Player

Fields

s_player :: Int

s_score :: Int

data Next

Informações sobre a pŕoxima jogada, nomeadamente o jogador que se segue e a peça que este deve colocar.

Constructors

Next

Fields

n_player :: Int

n_tile :: TileType

data Meeple

Um Meeple, com informação sobre o seu dono e tipo. O tipo do Meeple (F - Farmer, K - Knight ou M - Monk) define o sítio onde este está colocado num Tile.

Constructors

Meeple

Fields

m_player :: Int

m_type :: Char

data Tile

Uma peça de Terreno.

Constructors

Tile

Fields

t_type :: TileType

O tipo da peça.

t_x :: Int

Posição horizontal da peça.

t_y :: Int

Posição vertical da peça.

t_orientation :: Char

Orientação da peça (N,S,W,E).

t_meeple :: Maybe Meeple

Um Meeple que exista na peça.

data Board

O terreno.

Constructors

Board

Fields

b_terrain :: [Tile]

As peças do terreno.

b_scores :: [Player]

Os jogadores.

b_next :: Next

A próxima Jogada.

type TileType = Char

Um caracter (B,C,E ou N) que identifica o tipo da peça.

type Map = [[Maybe Tile]]

Uma matriz de peças para representar um mapa.

type Location = (Int, Int)

Um ponto de coordenadas (x,y).

data Limits

Os limites do mapa.

Constructors

Limits

Fields

l_Xmin :: Int

l_Xmax :: Int

l_Ymin :: Int

l_Ymax :: Int

getElemsFromContent :: [Content] -> [Element]

Obtém apenas os Element de um conjunto de Content.

getAttrValueInt :: String -> [Attr] -> Int

Encontra o valor inteiro para determinada chave num conjunto de atributos.

getAttrValueChar :: String -> [Attr] -> Char

Encontra o valor char para determinada chave num conjunto de atributos.

processaBoard :: Element -> Board

Encontra a tag <board> e processa os seus descendentes.

processaTerrain :: [Element] -> [Tile]

Encontra a tag <terrain> e processa os seus descendentes.

processaTile :: [Element] -> [Tile]

Encontra a tag <title> e processa os seus atributos e descendentes.

processaMeeple :: [Element] -> Maybe Meeple

Tenta encontrar uma tag <follower> e processa os seus atributos. Caso encontre constroi um Meeple.

processaScores :: [Element] -> [Player]

Encontra a tag <scores> e processa os seus descendentes.

processaScore :: [Element] -> [Player]

Encontra as tags <score>, processa os seus atributos e constroi uma lista de Player.

processaNext :: [Element] -> Next

Encontra a tag <next> e processa os seus atributos num Next.