Pontuar

Description

Trata de contar os pontos

Documentation

type ScoredTile = (Zone, Int)

Representa uma Zone e a sua pontuação

finalScore :: Board -> Board

Calcula as pontuações, inclusive dos Farmers, retira todos os meeples e atribui pontuações finais

updateScore :: Board -> Board

Retira os meeples e atribui pontuações

scorePlayers :: [ScoredTile] -> [Player] -> [Player]

Alterar os scores dos jogadores

removeMeeples :: [ScoredTile] -> [Tile] -> [Tile]

Produz uma nova lista de Tiles sem os Meeples que já foram pontuados

obterMeeplesPorRetirar :: Board -> [ScoredTile]

Obtém ScoredTiles que devem ser pontuados nesta jogada

getScoredTileFromTile :: Board -> Tile -> ScoredTile

Obtém um ScoredTile a partir de um Tile. A pontuação inicial é 0.

scoreZone :: Zone -> Int

Pontuar uma Zone

isMonkDone :: Board -> Tile -> Bool

Verifica se o Meeple monk deve ser retirado do tabuleiro

isKnightDone :: Board -> Tile -> Bool

Verifica se o Meeple knight deve ser retirado do tabuleiro

getRealScoresForZones :: [ScoredTile] -> [ScoredTile]

Obtém o verdadeiro score associado a uma zona depois de contar o número de Meeples presente nessa zona

joinScoredTilesByZone :: [ScoredTile] -> [ScoredTile] -> [[ScoredTile]]

Agrupa os ScoredTile caso se refiram à mesma Zone

playersToScore :: [ScoredTile] -> [Int]

Obtém uma lista com os jogadores que devem receber pontuação de acordo com o número de Meeples que têm numa Zone