Tabuleiro

Description

Contém funções para alterar e obter diversas informações sobre o tabuleiro

Documentation

data Zone

Representa uma zona do mapa: uma cidade, um campo ou um claustro, começando num determinado Tile. | O Tile inicial também deve estar na lista.

Constructors

City Tile [Tile]

Field Tile [Tile]

Cloister Tile [Tile]

nPecasB :: Int

Números de peças do tipo B que podem ser jogadas

nPecasC :: Int

Números de peças do tipo C que podem ser jogadas

nPecasE :: Int

Números de peças do tipo E que podem ser jogadas

nPecasN :: Int

Números de peças do tipo N que podem ser jogadas

nPecasT :: Int

Números de peças que podem ser jogadas (no total)

data Side

Representa a lateral de uma peça

Constructors

SideField

SideCity

SideBoth

type Sides = (Side, Side, Side, Side)

Representa as quatro laterais de uma peça

removeDuplicates :: Eq a => [a] -> [a]

Remove duplicados de uma lista

getTileAtLocation :: [Tile] -> Location -> Maybe Tile

Obtém o tile que está numa posição, ou Nothing caso não exista nenhum nessa posição

getLimits :: Board -> Limits

Obtém os valores máximos e mínimos para as coordenadas X e Y

getTileLimit :: (Tile -> Int) -> [Tile] -> (Int, Int)

Obtém o valor máximo para uma das coordenadas X ou Y, especificadas como t_x ou t_y

getSidesFromTile :: Tile -> Sides

Recebe um Tile e uma orientação e retorna as suas laterais (N,S,W,E)

getSidesFromTypeAndOrientation :: TileType -> Char -> Sides

Recebe um tipo de peça e uma orientação e retorna as suas laterais (N,S,W,E)

possibleBordersOfTileAt :: Int -> Int -> [Tile] -> Sides

Informa que tipo de tile pode estar na posição (x,y)

borderTypeMaybeTile :: Maybe Tile -> Sides

Dado um Tile obtém as suas bordas

northBorder :: Sides -> Side

Dado os lados, retornar o lado de cima

eastBorder :: Sides -> Side

Dado os lados, retornar o lado da direita

southBorder :: Sides -> Side

Dado os lados, retornar o lado de baixo

westBorder :: Sides -> Side

Dado os lados, retornar o lado da esquerda

removeAmbiguousSides :: Sides -> [Sides]

Desdobra Sides com lados ambíguos (SideBoth) numa lista de Sides que apenas são Field ou SideCity

getTilesFromMaybeTiles :: [Maybe Tile] -> [Tile]

Obtém uma lista de Tile que podem ser colocados numa posição a partir de uma lista de Sides

getTilesFromSides :: Int -> Int -> [Sides] -> [Maybe Tile]

Obtém o Tile que se pode colocar tendo em conta determinado Sides

getTilesFromSide :: Int -> Int -> Sides -> [Maybe Tile]

Verifica qual o tile que é possível colocar numa posição sabendo as laterais das peças adjacentes

possibleTilesAt :: [Tile] -> Location -> [Tile]

Dado uma localização e a lista de Tile em jogo, informa que peças podem ser lá colocadas

possibleNextTiles :: Board -> [Tile]

A partir do Board descobre todas as peças que podem ser colocadas, onde e em que posição (não verifica meeples)

getNextTileType :: Board -> TileType

Obter o tipo do próximo tile

getNextPlayer :: Board -> Player

Obter o próximo jogador

nextPlayerHasMeeples :: Board -> Bool

Verifica se o próximo jogador tem meeples por jogar

meeplesPlacedByNextPlayer :: Board -> Int

Verifica quantos meeples o próximo jogador tem em jogo

generateNext :: Int -> Board -> Next

Gera um elemento Next para a próxima jogada

generateNextPlayer :: Board -> Int

Determina o número do próximo jogador

randomValue :: Int -> Int -> Int

Fornecendo uma seed entre 0 e 1000, produz um número entre 0 e o segundo argumento.

pseudoRandomValue :: [Tile] -> Int -> Int

Gera um valor pseudo-aleatório no intervalo [0; maximo]

getTilesWithMeeples :: Board -> [Tile] -> [Tile]

A partir do tabuleiro e da lista de Tiles que podem ser colocados, devolver uma lista de Tiles que podem ser colocados com e sem Meepless

getTileWithKnight :: Board -> Tile -> Maybe Tile

Verificar todos os tile de cidade à volta a procura de um outro meeple, se nao existir, pode colocar

getTileWithFarmer :: Board -> Tile -> Maybe Tile

Verificar todos os tile de field à volta a procura de um outro meeple, se nao existir, pode colocar

getTileWithMonk :: Tile -> Maybe Tile

Verificar todos os tile de field à volta a procura de um outro meeple, se nao existir, pode colocar

canThisTileHaveMeeple :: Tile -> Board -> Char -> Bool

Verifica se um tile pode ter um meeple do tipo definido

searchMeeple :: Char -> [Tile] -> [Tile] -> [Tile] -> Bool

Procurar Meeples

tilesToAddBasedOnMeepleType :: Char -> Tile -> [Tile] -> [Tile]

Encontrar os tiles adjacentes que podem ter meeples do mesmo tipo

getFollowingTilesPositions :: Tile -> Side -> Sides -> [Location]

Obtém uma lista das posições dos Tiles por onde se deve continuar a procurar Meeples

isCityComplete :: [Tile] -> [Tile] -> [Tile] -> Bool

verifica se uma cidade está completa

pecasRestantes :: Board -> (Int, Int, Int, Int)

Verifica quantas peças de cada tipo (B,C,E,N) estão por colocar

contadorPecas :: [Tile] -> (Int, Int, Int, Int)

Verifica quantas peças de cada tipo (B,C,E,N) estão colocadas

randomValidTileToPlay :: Int -> Board -> Tile

Obter um tile aletório para jogar

playFirstTile :: Int -> Board -> Tile

Jogar a primeira peça

listOfTypesFromTuple :: (Int, Int, Int, Int) -> [TileType]

Obter uma lista de TileType a partir de um tuplo com as quantidades de peças (B,C,E,N)

isGameOver :: Board -> Bool

Verifica se o jogo terminou

substituteNext :: Board -> Next -> Board

Substitui a componente Next de um Board

getAllTilesWithMeeples :: Board -> [Tile]

Obtém todos os Tiles que têm Meeples

getFirstTile :: Zone -> Tile

Obtém o Tile inicial de uma Zone

isSameZone :: Zone -> Zone -> Bool

Verifica se duas Zones se referem à mesma região

tilesListsAreTheSame :: [Tile] -> [Tile] -> Bool

Verifica se duas lista de Tiles contêm os mesmos elementos

tilesAreTheSame :: Tile -> Tile -> Bool

Verifica se dois Tiles têm a mesma localização

getZoneForMeeple :: Board -> Tile -> Zone

Recebe um Tile com Meeple e obtém a zona dominada por esse Meeple

getCityTiles :: Tile -> [Tile] -> [Tile]

Obtém os Tiles de uma cidade começando no Tile especificado

getFieldTiles :: Tile -> [Tile] -> [Tile]

Obtém os Tiles de um campo começando no Tile especificado

getCloisterTiles :: Tile -> [Tile] -> [Tile]

obtém os Tiles junto ao claustro e o Tile do próprio claustro